+++ ChangeLog.md
@@ -4,9 +4,12 @@
 
 ### 0.4.0 (??)
 
+- add support for draw order in object layer
 - add support for Tiled 0.11.0 features
   - id for objects
   - hexagonal orientation
+- add support for Tiled 0.13.0 features
+  - tilecount in tileset
 - fix a bug in the parsing of the rotation attribute